\name{print.comparison}
\alias{print.comparison}
\title{
  Prints a Comparison Object.
}
\description{
  Prints the consensus comparison object. One line for each selected gconsensus method.
}
\usage{
\method{print}{comparison}(x, ...)
}
\arguments{
  \item{x}{comparison, object to be described.}
  \item{\dots}{any additional parameters.}
}
\value{
  A displayed text containing a summary of the selected gconsensus methods.
}
\author{
\packageAuthor{gconsensus}

Maintainer: \packageMaintainer{gconsensus}
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
  \code{\link{comparison.gconsensus}}, \code{\link{plot.comparison}}, \code{\link{toString.comparison}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ print }
\keyword{ comparison } % __ONLY ONE__ keyword per line

