\name{print.doe}
\alias{print.doe}
\title{
  Prints a Doe Object.
}
\description{
  Prints a doe (unilateral degrees of equivalence) object.
}
\usage{
\method{print}{doe}(x, ...)
}
\arguments{
  \item{x}{doe, object to be described.}
  \item{\dots}{any additional parameters.}
}

\value{
  A displayed text containg a detailed description of the object.
}
\author{
\packageAuthor{gconsensus}

Maintainer: \packageMaintainer{gconsensus}
}
\seealso{
  \code{\link{gconsensus}}, \code{\link{doe.gconsensus}}, \code{\link{toString.doe}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ print }
\keyword{ doe.gconsensus } % __ONLY ONE__ keyword per line
\keyword{ gconsensus } % __ONLY ONE__ keyword per line
