\name{print.gconsensus}
\alias{print.gconsensus}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Prints a gConsensus Object.
}
\description{
  Prints a gconsensus object.
}
\usage{
\method{print}{gconsensus}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{a gconsensus object.}
  \item{\dots}{any additional parameter.}
}
\details{
  option \code{digits} is used.
}
\value{
  A displayed text containing a detailed description of the object.
}
\author{
\packageAuthor{gconsensus}

Maintainer: \packageMaintainer{gconsensus}
}

\seealso{
  \code{\link{gconsensus}}, \code{\link{plot.gconsensus}}, \code{\link{toString.gconsensus}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ gconsensus }% use one of  RShowDoc("KEYWORDS")
\keyword{ plot.gconsensus }% __ONLY ONE__ keyword per line
