\name{toString.doe}
\alias{toString.doe}
\title{
  Builds a detailed description string of the doe object.
}
\description{
  This function converts a unilateral degrees of equivalence object into a string. The returned value contains a verbatim description of the object. This function is used to display the result of estimating the unilateral degrees of equivalence.
}
\usage{
\method{toString}{doe}(x, ...)
}
\arguments{
  \item{x}{doe, object to be described.}
  \item{\dots}{any additional parameters.}
}
\value{
  string, it contains a verbatim description of the unilateral degrees of equivalence object.
}
\author{
\packageAuthor{gconsensus}

Maintainer: \packageMaintainer{gconsensus}
}
\seealso{
  \code{\link{doe.gconsensus}}, \code{\link{print.doe}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ doe.gconsensus } % __ONLY ONE__ keyword per line
\keyword{ toString } % __ONLY ONE__ keyword per line
